+++ x/stake/genesis.go
@@ -1,6 +1,10 @@
 package stake
 
-import sdk "github.com/cosmos/cosmos-sdk/types"
+import (
+	tmtypes "github.com/tendermint/tendermint/types"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
 
 // GenesisState - all staking state that must be provided at genesis
 type GenesisState struct {
@@ -62,4 +66,17 @@ func WriteGenesis(ctx sdk.Context, k Keeper) GenesisState {
 		validators,
 		bonds,
 	}
+}
+
+// WriteValidators - output current validator set
+func WriteValidators(ctx sdk.Context, k Keeper) (vals []tmtypes.GenesisValidator) {
+	k.IterateValidatorsBonded(ctx, func(_ int64, validator sdk.Validator) (stop bool) {
+		vals = append(vals, tmtypes.GenesisValidator{
+			PubKey: validator.GetPubKey(),
+			Power:  validator.GetPower().Evaluate(),
+			Name:   validator.GetMoniker(),
+		})
+		return false
+	})
+	return
 }